Change `isinstance` check to fix incompatibility with Spark Connect (#24)
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary -->

In Spark 3.x, the Spark Connect's version of the `Column` isn't
inherited from `pyspark.sql.Column`, and as result, it's not recognized,
leading to generation of incorrect text. Instead of checking for
`Column` instance, we now check if the `message` is string or not.

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #23

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [x] manually tested
- [ ] added unit tests
- [ ] added integration tests

Signed-off-by: Alex Ott <alexey.ott@databricks.com>
Co-authored-by: Marcin Wojtyczka <marcin.wojtyczka@databricks.com>